requirements.txt
pika==0.12.0
pyyaml==3.13
